requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-genai==1.0.0
chromadb==0.4.24
numpy<2.0
sentence-transformers==3.0.0
rank-bm25==0.2.2
pydantic==2.9.0
python-multipart==0.0.9
pymupdf==1.24.0
python-docx==1.1.0
httpx==0.27.0
python-dotenv==1.0.0
aiofiles>=23.0.0
